UniversalRobot - Installation.mounting, .home, .tcp, .payload, and safety settings password
i edited the mounting section of installation so that the xy plane was oriented facing the tank. exactly. i set a home position that it could use for now. i used the wizards to calculate the tcp and the payload. and i created a safety settings password so that i could go in and fix a warning in there