Test Square Motion - a new program that moves in a square
it works. it does move in a square (or rectangle or whatever). if the pendant is in manual mode - create a program screen, it only runs when you are holding the button. but it automatic mode, it runs until you tell it to stop (no button holding needed). it must have repeat continuously checked or something.