Test Teach Position - get_actual_joint_positions()
with get_actual_tcp_pose() it was bending over backwards to get to the tcp position. not only was it taking a weird way (rotating vertically), but also the joints may be closer to their joint limits going that route, and we want to avoid encountering those joint errors as much as we can.