main program, move_path_1
main program now calls analyze_touchpoints.script, and also calls move_path_1, which moves the tcp along the weld path. it currently works until it tries to go around the corner. p2 is an inch off it seems. next week i want to modify this so that after it calculates the points, it displays a popup of the coordinates of the points. for troubleshooting