analyze_touchpoints script - get_corner_homing_45_pathpoints() add vertical touchpoints to check for nonvertical walls
i put in the touchpoints so they are calculated. want to angle the first path down the box so theres no gap at the top of that weld due to the box face angling away. still need to make the corner homing do it and save those points. still need to use those points to change the calculations of the weld points.