Corner Homing 45, Main Program RIght - incorporate vert touchpoints
Corner Homing 45 is now the one that incorporates the vertical touchpoints. Corner Homing 45 novert is the saved version of the one that does not. I tested Main Program right and it works well. the p1 is slightly forward (x-) and right (y+) of p2 on my test box. that is the change i was looking to see. I will have to test it tomorrow with a hot welder to see if it actually helps the weld quality.