move_path_3_left - added in all the stuff
move_path_3_left now has all the commands it needs in order to run. i'm not going to make a Main Program Left. I'm going to save Main_Program_Right as Main_Program. and then i'm going to add move_path_3_left in as a subprog. and then i'm going to add in a choice variable from Choice and depending on if the user selects left or right, call main_calculate_[left/right] and move_path_3[left/right].