Main program, movepath3right, default.install, munsch_metal.install - metal mounting plate installed
We got the metal mounting plate installed. we called it munsch_metal.installation. and saved it as default. its not quite complete. we jerry rigged the plastic touchpoint to be able to fit for now. the block is biger on the back side, and we wrapped the cords in a sheath. so, we needed to change the approach from home to weld ready position. this was completed in align_ready_for_welding, which is implemented in the main program for a right side.