default.install, munsch_metal - fixed backofhole, tried pushing in by 2.5 mm via touchpoint z change
the third leg was way off. i corrected the backofhole tcp. now it lines up with the second leg. still extruding out in space. so i forced it in 2.5 mm. now it is too far to the right on the first leg. still seems like it is extruding out not enough goop.