installation, mainprogram - move home to above cup
wade put a cup on the stand to be the home position for the welder. so i changed the home to be that. i also went through the main program and added a waypoint for the old home so that the path would go from new home to old home to next waypoint. this way the path from old home to next waypoint would be exactly the same.