MainProgram, CornerHoming45 - include sidewall variable reference, use moveP in between homing points
use moveP in between homing points to back up before going to the next point. going straight to the next homing point (with moveJ) can scrape the side of the tank or bonk a feature. with features potentially in the way, we need to back up before going over. moveP does this the quickest and smoothest (can avoid a pause in between each point).