MainProgram - switch back to one way communication on welds
switch the turn_extrusion_on() to suppressed, 6 times. switch the uR1 and Wait() to unsuppressed, 6 times. also change the speed of the around (moveL to p_2 and moveP) back to their original speeds. we were getting leaks, and hypothesized that it might have been the higher speed of the turn. this will check that.