Floor Corner Extended - switch corner homing retreats to moveLs, switch weld approaches to moveLs
moveP on the retreat sometimes fails. switched to moveL so it doesnt need constant speed. i should switch it to a standard moveJ though. would be faster. weld approach was too fast, so i slowed them down with moveLs. this latest still failed on point_5 on the right side weld.