analyze touchpoints - get_corner_homing_45_pathpoints() - add pointtop param
when deciding on homing points, if it's an extended weld, push the touchpoints up to the start of the weld. points farther apart on the line give you a more accurate line for the duration of the line. do this to prevent missing the mark at the top of the weld.