analyze touchpoints script - include chamferup in the extended functions
there have been multiple tanks with 3/4 inch floors come through, that the robot would have been able to weld but couldn't. i just tried to weld a corner and it laid a bad weld because of this. i didn't recognize it. having the popup question will also remind me to check. we gain access to some corners and its a simple click.