analyze touchpoints script - remove left around extension
sometimes the floor of the tank was not level (or maybe the robot cart was not level) but sometimes on the extension, the weld got off track of the chamfer and laid a bad weld. I don't want to add touchpoints or change this touchpoint location so ill shorten it back to normal to reduce distance for it to get off track. I don't want to keep having to watch it do the around and stop it manually when it gets off track.