analyze touchpoints script - add floor_corner_lift_column_extend_right()
the left side auto-extending is working great. now add that functionality to the right side. I don't have all the angles so for now I'll just skip that check and just assume the bottom points will be good. This should work. but it doesnt (at least not when i put it into the program). after the first turn, it hit 1.5 cm above the floor chamfer