analyze touchpoints script - adjust minimum height
cameron was running the station and he likes the tank on the blocks at the lowest level. when putting tcp at the corner, program said it was ~ 1 inch too low. however, when we ran the program, we had multiple inches to spare. we will try this and see if it works. hopefully it won't bonk, even if the cords prush the ground slightly that could be okay