Robot2 initial backup commit
I did a backup of robot2. a bunch of stuff is changed. i chose 'Discard Changes' for all the files that it was missing. This commit includes a bunch of programs of me messing around with waypoints trying to get the step corner path figured out. It also includes changes that were made to subprograms a while ago inside Floor Corner Extended, but never saved out to the actual subprogram files. It also includes some useful programs like freedrive_mode1