Robot 2 step corner update
the right side step corner is working in the program Step Corner.urp EXCEPT for the circle move. circle move is still pretty clunky. it is too fast. i slowed the speed down to 0.2 and it is still too fast. probably because of a math error making it not good coplanar. because when it gets to point 4 the orientation is off as well. probably because heading1 and heading3 aren't exactly orthogonal.