analyze touchpoints script - cover corners - deepchamfers
if the covers have deep chamfers, we need to take that into account by bumping the point over farther into the chamfer so it's centered, and slowing the welder speed down. we can't bump point2 over because the vertical seam is still small, and it needs to come back to regular chamfer size by alignment by the time it gets to the v chisel out of the endwall. point2a can't be bumped over because the path from p2a to p2 needs to be in the same plane as the circle move.