analyze touchpoints - floor corner - reduce weldistance 2 and 3
that's not the focus. just weld the corner and the vert str. the weldpath 3 was getting too thin at the end anyways. due to larger chamfers in florida. i dont really want to slow down the speed, because that might mess up the around. but just decreasing the distance will prevent that weld thinning.