analyze touchpoints - floor corner extended - increase maximum z of ph2 and ph4
for setting the homing startpoints, increase the maximum z of the vert points. before, the maximum z was only getting like halfway up the tall florida tanks, and the bowing was making it come in inaccurate when it went to point1 to weld. not sitting in chamfer. increasing the top point might make the homing more accurate near the top.