analyze touchpoints - floor corner - get_corner_homing_45_pathpoints() - add middle vert points
for floor corner welds, in florida the tank is really high. we fixed the probing accuracy at the very top, but it started coming out towards the bottom middle. to do this we will add a third touchpoint halfway between the first and second. this is the code to calculate that touchpoint. i will then need to program the robot to go to it and save a point. i will then need to pass those additional points into the main function. i will then need to adjust analyze_vertical_touchpoints() to calculate the midpoint. i will then need to pass the midpoint of the weldpath back out. i will then need to adjust the weldpath program to actually go to that point before ph1 and ph2. but it's time to go home right now.