analyze_floor_corner.script - modify welddistances
even with a half inch difference, the first weld was going a half inch farther than the second weld. additionally, i had lucas weld the box according to the lines. and i v chiseled them out. when going into the tilt back, we had ample space to go further on the straight.